added navigation for buttons to open new "windows"
