+++ getaway/App.js
@@ -1,161 +1,73 @@
-import React, { useState } from 'react';
-import { StatusBar, StyleSheet, Text, View, TouchableOpacity, ImageBackground } from 'react-native';
+import React from 'react';
+import { Text, View, ImageBackground, StatusBar, StyleSheet, TouchableOpacity } from 'react-native';
+import { NavigationContainer } from '@react-navigation/native';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import Icon from 'react-native-vector-icons/Ionicons';
+
+import HomeScreen from './HomeScreen';
+
+
+
+function SearchScreen() {
+  return <Text>Search Screen</Text>;
+}
+
+function LikeScreen() {
+  return <Text>Like Screen</Text>;
+}
+
+function ProfileScreen() {
+  return <Text>Profile Screen</Text>;
+}
+
+const Tab = createBottomTabNavigator();
 
 export default function App() {
   return (
-    
-      <View style={styles.container}>
-        <ImageBackground source={require('./assets/pietro-de-grandi-pidpjXuorPE-unsplash.jpg')} style={styles.image}>
-        <View style={styles.takeATrip}>
-          <Text style={styles.titleName}>Get Away</Text>
-          <TripButtons/>
-        </View>
-        </ImageBackground>
-        <StatusBar style="auto" />
-        <View style={styles.line}/>
-        <ButtonHub/>
-      
-      </View>
-    
+    <View style={styles.container}>
+      <StatusBar style="auto" />
+
+      <NavigationContainer>
+        <Tab.Navigator
+          screenOptions={({ route }) => ({
+            tabBarIcon: ({ focused, color, size }) => {
+              let iconName;
+
+              if (route.name === 'Home') {
+                iconName = focused ? 'home' : 'home-outline';
+              } else if (route.name === 'Search') {
+                iconName = focused ? 'search' : 'search-outline';
+              } else if (route.name === 'Likes') {
+                iconName = focused ? 'heart' : 'heart-outline';
+              } else if (route.name === 'Profile') {
+                iconName = focused ? 'person-circle' : 'person-circle-outline';
+              }
+
+              return <Icon name={iconName} size={size} color={color} />;
+            },
+
+            tabBarActiveTintColor: '#FD9528',
+            tabBarInactiveTintColor: '#10384a',
+            tabBarStyle: { position: 'absolute' },
+            tabBarShowLabel: false,
+            headerShown: false,
+          })}
+          
+        >
+          <Tab.Screen name="Home" component={HomeScreen} />
+          <Tab.Screen name="Search" component={SearchScreen} />
+          <Tab.Screen name="Likes" component={LikeScreen} />
+          <Tab.Screen name="Profile" component={ProfileScreen} />
+        </Tab.Navigator>
+      </NavigationContainer>
+    </View>
   );
 }
-// Button Hub
-const ButtonHub = () => {
-  const [homePressed, setHomePressed] = useState(false);
-  const [searchPressed, setSearchPressed] = useState(false);
-  const [likePressed, setLikePressed] = useState(false);
-  const [profilePressed, setProfilePressed] = useState(false);
-  return (
-  <View style={styles.hub}>
-    <View style={styles.button}>
-    <Icon.Button 
-      name="home"
-      backgroundColor="white"
-      size={30}
-      color={homePressed ? '#FD9528' : '#10384a'}
-      onPress={() => {
-        console.log('Home Button Pressed');
-        setHomePressed(!homePressed); // Toggle the pressed state
-      }}
-    />
-    </View>
-    
-    <View style={styles.button}>
-    <Icon.Button
-      name="search"
-      backgroundColor="white"
-      size={30}
-      color={searchPressed ? '#FD9528' : '#10384a'}
-      onPress={() => {
-        console.log("Search Pressed");
-        setSearchPressed(!searchPressed);
-      }}
-    />
-    </View>
-    <View style={styles.button}>
-    <Icon.Button
-      name="heart"
-      backgroundColor="white"
-      size={30}
-      color={likePressed ? '#FD9528' : '#10384a'}
-      onPress={() => {
-        console.log('Like button Pressed');
-        setLikePressed(!likePressed);
-      }}/>
-    </View>
-    <View style={styles.button}>
-    <Icon.Button
-      name="person-circle-outline"
-      backgroundColor="white"
-      size={30}
-      color={profilePressed ? '#FD9528' : '#10384a'}
-      onPress={() => {
-        console.log("Profile Button Pressed");
-        setProfilePressed(!profilePressed);
-        }}
-      />
-    </View>
-    
-  </View>
-  );
-};
-const TripButtons = () => (
-  <View style={styles.tripButtons}>
-    <TouchableOpacity style={styles.topButton}>
-      <Text style={styles.topButtonText}>Generate Trip</Text>
-    </TouchableOpacity>
-    <TouchableOpacity style={styles.topButton}>
-      <Text style={styles.topButtonText}>Plan Route</Text>
-    </TouchableOpacity>
-  </View>
-);
+
 const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: '#fff',
-    alignItems: 'center',
     justifyContent: 'center',
-  },
-  hub: {
-    flexDirection: 'row',
-    bottom: 40,
-    position: 'absolute',
-    width: '100%'
-  },
-  button: {
-    flex: 1,
-    alignItems: 'center',
-    
-  },
-  titleName: {
-    fontSize: 55,
-    alignSelf: 'center',
-    paddingTop: 15,
-    color: '#002527'
-  },
-  line: {
-    position: 'absolute',
-    bottom: 90,
-    width: '100%',
-    borderTopWidth: 1,
-    borderColor: 'grey',
-  },
-  takeATrip: {
-    height: '25%',
-    width: '80%',
-    borderWidth: 1,
-    borderColor: 'grey',
-    marginTop: -450,
-    alignSelf: 'center',
-    borderRadius: 10,
-    overflow: 'hidden',
-  },
-  tripButtons: {
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-    padding: 10,
-    paddingBottom: 25,
-    width: '100%',
-    height: '100%',
-    alignItems: 'flex-end',
-    position: 'absolute',
-  },
-  topButton: {
-    backgroundColor: '#FD9528',
-    width: 110,
-    height: 60,
-    borderRadius: 10,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  topButtonText: {
-    color: 'white',
-  },
-  image: {
-    width: '100%',
-    height: '100%',
-    justifyContent: "center",
-    paddingBottom: 91,
   },
 });
